chore(setup): initialize phoenix application
The phoenix application has been generated using the `phoenix.new` mix
task. The `no-brunch` switch was passed to this task as it is not
anticipated that we will be using the asset pipeline for this project.

A `config/env.secret.exs.example` has been included, with the purpose of
providing a template for environment specific config overrides. It is
anticipated that developers will create a copy of this file for each of
their environments (dev, test, prod):

```bash
cp config/env.secret.exs.example config/dev.secret.exs
```

A gitignore rule has been setup to ignore files in the `config`
directory that end in `secret.exs`.

The `config/config.exs` file has been modified to import the relevant
environment overrides should the file exist.