feat(User): add user model with basic GET route
A new `Thumbifier.User` model has been added that holds the relevant
account information for access to the API.

The GET route is at `/users/:email` and will return a JSON encoded
string of the found user information. It will return a JSON encoded
error if no user matches the given email.

There is no authentication for getting user details at the moment, but
this will be implemented in the future.

A `Util.Time` module has been added to handle basic DateTime functions
that Ecto.DateTime doesn't support, it utilises the Timex package.