feat(Plug.Authenticated): add plug for authenticating user requests
To obtain user details, it is now required to provide the email in the
GET request `/users/:email` and the api_token associated with that email
address passed as a `Authorization: Bearer <secret>` header.

If the given email address is authenticated with the provided api_token
then the user's details are returned successfully, otherwise a status
code of 401 is returned.

This will ensure that only an authenticated user will be able to access
the api_grant for accessing the API functionality.

The `api_token` will be SHA512 hashed when stored in the database so
when the User is queried using the `api_token` it will be hashed
accordingly prior to the lookup.

Due to this addition, it is no longer possible to fail an email address
match in the controller, so the test has been removed, however the
controller code remains as a fallback check.