feat(ApiGrant): generate unique `api_grant`
To access the API calls, a unique  ApiGrant is generated and returned
when the user details are requested via the GET request on the
`/users/:email` route.

The user's email is passed to the creation of the ApiGrant so the
ApiGrant can be loosely associated with the originating user.

The ApiGrant is returned as part of the JSON from the GET route for
the User.