feat(User): add POST route to create new user
The POST route of `/users/` will accept an `email` parameter and create
a new user if the email is valid and not a duplicate.

On success, the route will return a JSON response containing the email
address given and the api_token associated with the user account.

On failure, the route will return a 422 status code and a JSON response
containing the errors with the request.