feat(User): add PUT route for updating user email
The PUT route of `/users/:email` will accept a `new_email` parameter and
update the user's email address if the new email is valid and not a
duplicate.

On success, the route will return a JSON response containing the
previous email address and the newly updated current email address.

On failure, the route will return a 422 status code and a JSON response
containing the errors with the request.