feat(URI): ability to obtain details of a uri
A new `Thumbifier.Util.URI` module has been added that provides functionality for parsing
and processing a URI.

The `valid?/1` function returns `true` if the given URI is well-formed
and returns `false` if it isn't. This function replaces
`Thumbifier.Util.Validation.uri` as this previously provided URI
validation, but as it has now been superseded the
`Thumbifier.Util.Validation` module has now been deleted.

The `Thumbifier.Util.URI` module also provides a function to obtain
details of a given URI. it returns the following:

  * A file with a size of 1234bytes URI is given `{ok: [size: 1234, type: :file]}`
  * A website URI is given `{ok: [size: 0, type: :site]}`
  * An error when an invalid URI is given `{error: "<given uri> is not a
  * valid URI"}`

Another helper module has been added to make shell calls
`Thumbifier.Util.Shell` and this utilises `devinus/sh` to execute shell
commands. The reason this has been extracted into a helper module is
because it makes it easier to mock the shell calls as `Sh` seems
difficult to mock.