feat(Processor): implement worker processor to convert pdf to jpg
The Processor module now takes the POSTed form data passed to the
PoolBoy worker and calls the relevant functions to convert the given
`media_url` to a jpg and to POST the base64 encoded output to the given
`callback_url`.

The following form fields are POSTed to the `callback_url`:

  * status => `success` || `error`
  * response_id => the unique ID returned in the original POST
  * personal_reference => the reference given in the original POST
  * payload => base64 encoded image || error message

Also, the test remote PDF file URL has been updated to point to the one
in the github repository rather than rely on a 3rd party website.