refactor(Processor): generate expected image output `on-the-fly`
Previously the expected image output was asserted against a fixture file
by comparing the base64 encoded string of the fixture file and the image
file generated by the Processor.

This is too inconsistent as the version of ImageMagick will generate a
different file structure to the expected image output and therefore
there is a chance the base64 encoded version of the files may not match
even though the images look the same.

This solution uses the local fixture file to generate the expected
output to ensure the assertion is always identical.

Whilst, it doesn't prove the image is generated correctly as it will
generate the same base64 signature even on error, it at least determines
if the output sent to the callback url is well formed.

Testing the actual image output should be done via an end-to-end test
and using image comparison (such as ImageMagick compare function) to
ensure visual accuracy.