refactor(Processor): return payload from process function to worker
Previously, the worker module would simply call the process function and
all of the inner workings was encapsulated, which made testing the
actual image output very difficult, especially when different version of
ImageMagick can generate different base64 output for the same converted
image.

Now, the process function returns a payload containing the base64
encoded processed jpg and the worker module is responsible for POSTing a
payload to the callback-url.

This refactor means testing the process function is easier with
different versions of ImageMagick as now the generated jpg can be
compared with the test fixture using the ImageMagick `compare` function
as this will determine if the generated image is the same (or similar
enough to be considered the same) as the expected image stored in the
test/fixtures.