chore(Auth): rename authentication/authorisation items
Previously the authentication for a user was called an `api_token` which
could cause confusion as tokens can also be used for authorisation such
as with OAuth. So the authentication item is now called the `api_key` as
this is the generated unique key for the account.

The authorisation was previously referred to as the `api_grant` but
again this can cause confusion with other things such as OAuth so to
keep it simple it has been renamed to the `access_token`. The
`access_token` is the item that is provided for accessing the API
functions and is obtained with the user's email and their unique
`api_key`