fix(User): ensure usage recounter is reset after expiry time
Previously, the User `usage_cycle_expired?` function was relying on the
comparison of `Ecto.DateTime` structs which are not comparable as the
order of the data in the struct is not guaranteed.

This means it's not possible to perform a basic comparison on
`Ecto.DateTime` and instead the values are converted to Erlang tuples
where the order of the data is guaranteed and therefore comparable.