fix(URI): ensure wget does not crash process on file not found
If the `Sh` module receives a non-zero return from a command line
process then it raises an exception which causes the `GenServer` to
terminate.

This can happen if when calling `Sh.wget` a URL is provided that does
not exist (but is well formed) as wget cannot find the file to query and
exits with non-zero causing `Sh` to raise the exception.

The solution at present is to rescue from the exception when calling
`Sh.wget` and returning an appropriate payload that the `Util.URI`
module can process and provide a more appropriate response (such as
returning an error that the file was not found).